move base classes to base.py
