+++ ord_schema/orm/structure.py
@@ -23,7 +23,7 @@ from sqlalchemy import Column, Index, Integer, ForeignKey, Text, func
 from sqlalchemy.orm import with_polymorphic
 from sqlalchemy.types import UserDefinedType
 
-from ord_schema.orm import Base, Child, Parent
+from ord_schema.orm.base import Base, Child, Parent
 
 
 def rdkit_cartridge() -> bool:
